PHPStan Level 7: Just annoying. Fixed type issues

+++ src/xenialdan/MagicWE2/selection/shape/Cuboid.php
@@ -38,9 +38,9 @@ class Cuboid extends Shape
 
     public static function constructFromPositions(Vector3 $pos1, Vector3 $pos2): self
     {
-        $width = abs($pos1->getX() - $pos2->getX()) + 1;
-        $height = abs($pos1->getY() - $pos2->getY()) + 1;
-        $depth = abs($pos1->getZ() - $pos2->getZ()) + 1;
+        $width = (int)abs($pos1->getX() - $pos2->getX()) + 1;
+        $height = (int)abs($pos1->getY() - $pos2->getY()) + 1;
+        $depth = (int)abs($pos1->getZ() - $pos2->getZ()) + 1;
         $cuboid = new Cuboid((new Vector3(($pos1->x + $pos2->x) / 2, $pos1->y, ($pos1->z + $pos2->z) / 2)), $width, $height, $depth);
         return $cuboid;
     }
